Make Hit component render configurable
